Closes #35 Auth0 login, logout and form protection added.

+++ pages/api/deletePost.js
@@ -1,11 +1,20 @@
-import { deletePost } from '../../utils/fauna';
+import { deletePost, getPost } from '../../utils/fauna';
+import { withApiAuthRequired, getSession } from '@auth0/nextjs-auth0';
 
-export default async function handler(req, res) {
+export default withApiAuthRequired(async function handler(req, res) {
+  const session = getSession(req, res);
+  const userId = session.user.sub;
+
   if(req.method !== 'DELETE') {
     return res.status(405).json({ msg: 'Method not allowed.' });
   }
 
   const { id } = req.body;
+  const existingRecord = await getPost(id);
+  if(!existingRecord || existingRecord.data.userId !== userId) {
+    return res.status(404).json({ msg: 'Record not found.' });
+  }
+  
   try {
     const deletedPost = await deletePost(id);
     return res.status(200).json(deletedPost);
@@ -13,4 +22,4 @@ export default async function handler(req, res) {
     console.error(err);
     res.status(500).json({ msg: 'Something went wrong.' });
   }
-}
+});
